chore(git): add .gitattributes file to repository
Add a .gitattributes file to the repository, to assign attributes to
files and folders in the repository.

This file was taken from alexkaratarakis/gitattributes@5a965d3. This
specific .gitattributes file, amongst other things:
- auto-detects text files and performs LF normalization
- explicitly declares common files as either text and binary
- explicitly declares known OS-specific files as ending with CRLF or LF
- converts complex file types to text (doc, pdf, etc) before a diff
- ignore the .gitignore and .gitattribute files when exporting the repo

A guide on the .gitignore file's capabilities can be found on David
Laing's blog, at https://bit.ly/36QX6UD

As that repository is licensed under MIT, the repo's license notice was
added to the top of the file as a comment.