feat(action): add support for semantic-release
Add a job in the GitHub CI workflow to automatically tag commits in the
main and develop branches, based on the commit message headers. As long
as the commit messages follow the Conventional Commits specification,
semantic-release should be able to identify them and create new releases
using Git tags.

Releases are only created on the main and develop branches. Releases in
the develop branch are marked as pre-release.

The workflow was also renamed to something more generic, since it no
longer exclusively runs pre-commit.

Associated issue: ShahradR/git-template#8