docs(license): update project license to MIT-0
Update the project to use the MIT No Attribution (MIT-0) license.
Because this project is a template intended to be re-used by other
people to build their own projects, it doesn't make much sense to use a
license that requires attribution.

To see more details around the license and the rationale behind it,
see:
- https://bit.ly/2QJSR7I for the MIT-0 SPDX entry
- https://bit.ly/3gQYOu5 for a rationale behind the license
- https://git.io/JUml1 for Amazon's use of the modified license