feat: add a Code of Conduct to the project
Adopt the Contributor Covenant (https://www.contributor-covenant.org/)
in this project, and add a CODE_OF_CONDUCT.md file to the root of the
repository.

The README file was also updated to include a badge indicating that this
project is governed by the Contributor Covenant, as well as a section
helping consumers of this project update the contact information in the
Code of Conduct when creating a new repository based on this template.

This commit also updates the .alexignore, .pre-commit-config.yaml, and
.vale.ini files to ignore the CODE_OF_CONDUCT file when running Alex and
Vale, since the terms of the Code of Conduct contains wording caught by
the two linters.