feat: add a CONTRIBUTING document to the project
Add the first draft of the CONTRIBUTING document to the project. This
document defines contributing guidelines for this project, and helps
participants understand how to collaborate within this community.

This initial draft details the commit message standards used in this
project, as well as information regarding this project's Code of
Conduct. Some details around the environment setup was also added and
will be expanded upon in a future commit.