feat: add markdownlint-cli2 to this project
Add support for markdownlint-cli2 to this project, "a fast, flexible,
configuration-based command-line interface for linting Markdown/
CommonMark files with the markdownlint library".

This commit also updates the Prettier rules to wrap lines in Markdown
files to 80 characters, and adds a markdownlint configuration file to
allow the "summary" and "details" HTML elements without triggering rule
MD033.

The markdownlint-cli2 tool is installed as a pre-commit hook and in the
development container.

See https://bit.ly/3PV7oJS for more details.