feat: add codespell to the project
Add codespell to the project, a tool to "check code for common
misspellings".

A configuration file is also added to ignore files in the vale/
directory since they are sourced from an external repository and
codespell generates false positives on some of the Vale configuration
files.

This tool is installed in the development container and as a pre-commit
hook. See https://bit.ly/3AuIOKo for more details.