docs: publish MkDoc documentation to GitHub Pages
Update the MkDocs configuration for hosting on GitHub Pages. Because
GitHub is deprecating the branch-style deployment, we want to avoid
using the `mkdocs gh-deploy` command described by https://bit.ly/3Tsske8

Instead, we configure MkDocs to deploy as a local files" style, as
described in the https://bit.ly/3eNiOTH guide. By running the `mkdocs
build` command in a GitHub workflow similar to the one described in
https://bit.ly/3gu11Bz we can deploy the GitHub Pages using
environments.