docs: enable admonitions in documentation
Enable admonitions in the Material for MkDocs pages and update existing
documentation to leverage them when rendering notes.

Note that the link-image-reference-definitions rule had to be disabled
when running markdownlint, since the indented text is recognized as a
code block by the CommonMark specification.

While markdown-it plugins are supported and could resolve the issue, the
existing markdown-it-admonitions supports a different syntax, and the
link reference is still recognized as code. See https://bit.ly/3sc2AXc
for more details.