style: allow script tags in Markdown docs
Allow script tags to be added in Markdown documents when linting using
markdownlint, preventing error MD0033 from being generated when the
element is used to create dynamic content for the MkDocs documentation.

Because this exception has only been created because we are using
MkDocs, the template cleanup workflow was also updated to remove this
exception when creating a new project using this template repository.