chore: ignore tags when linting MkDocs config
Because the MkDocs configuration file can sometimes include YAML tags,
the ruaruamel.yaml linter will fail when trying to resolve them. To
bypass this, we pass the `--unsafe` flag to the linter, instructing it
only to perform syntax checks.

Because we are reducing the linter's capabilities for an edge case, we
add another definition in pre-commit configuration, and only scan the
mkdocs.yml file with the `--unsafe` flag.

To ensure that this doesn't affect users cloning the repository, new
`sed` commands are included in the template cleanup workflow to remove
the modifications made to support the MkDocs configuration file.